// features/p2p/types.ts
export interface Challenger {
  playerId: string
  playerName: string
}